Listings and Users

+++ iatbdApp/routes/web.php
@@ -1,10 +1,10 @@
 <?php
 
-use Illuminate\Support\Facades\Route;
-use Illuminate\Http\Request;
-use App\Models\Listing;
 use App\Http\Controllers\ListingController;
 use App\Http\Controllers\UserController;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Route;
+use App\Models\Listing;
 
 /*
 |--------------------------------------------------------------------------
@@ -20,31 +20,31 @@ use App\Http\Controllers\UserController;
 Route::get('/', [ListingController::class, "index"]);
 
 // Show the create page
-Route::get('/listings/create', [ListingController::class, "create"]);
+Route::get('/listings/create', [ListingController::class, "create"])->middleware("auth");
 
 // Store new listing data
-Route::post("/listings", [ListingController::class, "store"]);
+Route::post("/listings", [ListingController::class, "store"])->middleware("auth");
 
 //Show Edit Form
-Route::get("listings/{listing}/edit", [ListingController::class, "edit"]);
+Route::get("listings/{listing}/edit", [ListingController::class, "edit"])->middleware("auth");
 
 // Submit Edit update
-Route::put("listings/{listing}", [ListingController::class, "update"]);
+Route::put("listings/{listing}", [ListingController::class, "update"])->middleware("auth");
 
 // Delete
-Route::delete("listings/{listing}", [ListingController::class, "destroy"]);
+Route::delete("listings/{listing}", [ListingController::class, "destroy"])->middleware("auth");
 
 // Show Register Form
-Route::get("/register", [UserController::class, "create"]);
+Route::get("/register", [UserController::class, "create"])->middleware("guest");
 
 //Create new userf
 Route::post("/users", [UserController::class, "store"]);
 
 // Logout
-Route::post("/logout", [UserController::class, "logout"]);
+Route::post("/logout", [UserController::class, "logout"])->middleware("auth");
 
 // Login
-Route::get("login", [UserController::class, "login"]);
+Route::get("login", [UserController::class, "login"])->name("login")->middleware("guest");
 
 // LoginUser
 Route::post("/users/authenticate", [UserController::class, "authenticate"]);
